member_a_search/requirements.txt
numpy
pandas
Pillow
faiss-cpu
sentence-transformers